Move Auth0ServiceProvider to related namespace/directory
