+++ config/app.php
@@ -191,8 +191,8 @@ return [
         /*
          * Application Service Providers...
          */
+        App\Auth0\Auth0ServiceProvider::class,
         App\Domain\Integrations\IntegrationServiceProvider::class,
-        App\Providers\Auth0ServiceProvider::class,
         App\Providers\AuthServiceProvider::class,
         // App\Providers\BroadcastServiceProvider::class,
         App\Providers\HorizonServiceProvider::class,
